+++ common/src/main/java/traben/entity_model_features/mixin/MixinResourceReload.java
@@ -4,6 +4,7 @@ import net.minecraft.client.MinecraftClient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import traben.entity_model_features.EMFClient;
 import traben.entity_model_features.utils.EMFManager;
@@ -13,9 +14,15 @@ import traben.entity_model_features.utils.EMFManager;
 public abstract class MixinResourceReload {
 
     @Inject(method = "reloadResources()Ljava/util/concurrent/CompletableFuture;", at = @At("HEAD"))
-    private void emf$reload(CallbackInfoReturnable<Float> cir) {
+    private void emf$reloadStart(CallbackInfoReturnable<Float> cir) {
         if (EMFClient.testForForgeLoadingError()) return;
         EMFManager.resetInstance();
+    }
+
+    @Inject(method = "onFinishedLoading", at = @At("HEAD"))
+    private void emf$reloadFinish(final CallbackInfo ci) {
+        if (EMFClient.testForForgeLoadingError()) return;
+        EMFManager.getInstance().modifyEBEIfRequired();
     }
 }